pyproject.toml
[project]
name = "ayy"
version = "0.1.3"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Hamza Farhan", email = "thehamza96@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.39.0",
    "burr[graphviz,hamilton,opentelemetry,start,tracking-client,tracking-server]>=0.32.0",
    "google-generativeai>=0.8.3",
    "instructor>=1.6.3",
    "loguru>=0.7.2",
    "moviepy>=1.0.3",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.22",
    "valkey>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.hatch.build.targets.sdist]
exclude = ["**/", "!src/"]
